DHSCFT-971: refiled the heatmap
