+++ frontend/fingertips-frontend/components/charts/HeatMap/components/cell/DataCell.styles.tsx
