+++ frontend/fingertips-frontend/components/charts/HeatMap/components/cell/IndicatorCell.styles.tsx
