+++ frontend/fingertips-frontend/components/charts/HeatMap/components/cell/IndicatorPeriodCell.tsx
