+++ frontend/fingertips-frontend/components/charts/HeatMap/components/cell/IndicatorTitleCell.tsx
