+++ frontend/fingertips-frontend/components/charts/HeatMap/components/header/AreaBenchmarkGroupHeader.tsx
