+++ frontend/fingertips-frontend/components/charts/HeatMap/components/header/AreaHeader.tsx
