+++ frontend/fingertips-frontend/components/charts/HeatMap/components/header/AreaNonBenchmarkGroupHeader.tsx
