+++ frontend/fingertips-frontend/components/charts/HeatMap/components/header/IndicatorHeader.styles.tsx
