+++ frontend/fingertips-frontend/components/charts/HeatMap/components/header/IndicatorPeriodHeader.tsx
