+++ frontend/fingertips-frontend/components/charts/HeatMap/components/hover/BenchmarkPill.styles.tsx
