+++ frontend/fingertips-frontend/components/charts/HeatMap/components/hover/BenchmarkPillIcon.test.tsx
