+++ frontend/fingertips-frontend/components/charts/HeatMap/components/hover/BenchmarkPillIcon.tsx
