+++ frontend/fingertips-frontend/components/charts/HeatMap/helpers/prepareHeatmapData.test.ts
